@@ -1,15 +1,7 @@
 #include "bioNonlinearTissue.h"
-#include "bioAnalysis.h"
-#include "bioNeoHookeanIntegrator.h"
-#include "bioTrnsIsoNeoHookeanIntegrator.h"
-#include "bioHolmesMowIntegrator.h"
-#include "bioVariableRecovery.h"
-#include <ErrorEstimators.h>
 #include <amsiControlService.h>
-#include <amsiLinearStressStrainIntegrator.h>
 #include <apfFunctions.h>
-#include <apfsimWrapper.h>
-#include <simBoundaryConditions.h>
+#include <apfLabelRegions.h>
 #include <cassert>
 #include <cstdio>
 #include <cstdlib>
@@ -22,15 +14,18 @@
 #include <stdexcept>
 #include <string>
 #include <vector>
+#include "bioAnalysis.h"
 #include "bioHolmesMowIntegrator.h"
+#include "bioNeoHookeanIntegrator.h"
 #include "bioTrnsIsoNeoHookeanIntegrator.h"
 #include "bioVariableRecovery.h"
+#include "gmi.h"
 namespace bio
 {
-  NonlinearTissue::NonlinearTissue(pGModel imdl, pParMesh imsh, pACase pd,
-                                   pACase ss, MPI_Comm cm)
-      : FEA(cm)
-      , apfSimFEA(imdl, imsh, pd, ss, cm)
+  NonlinearTissue::NonlinearTissue(apf::Mesh * mesh,
+                                   const mt::CategoryNode & analysis_case,
+                                   MPI_Comm cm)
+      : amsi::apfFEA(mesh, analysis_case, {}, {}, "macro", cm)
       , constitutives()
       , dv_prev(0.0)
       , load_step(0)
@@ -46,15 +41,19 @@ namespace bio
     apf::zeroField(delta_u);
     apf_primary_delta_field = delta_u;
     // create a current coordinate field from the CurrentCoordFunc (x = X+u)
-    xpyfnc = new amsi::XpYFunc(apf_mesh->getCoordinateField(),apf_primary_field);
+    xpyfnc =
+        new amsi::XpYFunc(apf_mesh->getCoordinateField(), apf_primary_field);
     // user field is not zeroed because it is the sum of two other fields
-    current_coords = apf::createUserField(apf_mesh, "current_coordinates",
-                                          apf::VECTOR, apf::getLagrange(1), xpyfnc);
-    // take the current coordinate field, and subtract increment in displacement to
-    // get the coords at the previous increment (this is sorta hacky but it should work)
+    current_coords =
+        apf::createUserField(apf_mesh, "current_coordinates", apf::VECTOR,
+                             apf::getLagrange(1), xpyfnc);
+    // take the current coordinate field, and subtract increment in displacement
+    // to get the coords at the previous increment (this is sorta hacky but it
+    // should work)
     prv_crd_fnc = new amsi::XpYFunc(current_coords, delta_u, 1, -1);
-    prev_coords = apf::createUserField(apf_mesh, "previous_coordinates",
-                                       apf::VECTOR, apf::getLagrange(1), prv_crd_fnc);
+    prev_coords =
+        apf::createUserField(apf_mesh, "previous_coordinates", apf::VECTOR,
+                             apf::getLagrange(1), prv_crd_fnc);
     apf_primary_numbering = apf::createNumbering(apf_primary_field);
     strs = apf::createIPField(apf_mesh, "stress", apf::MATRIX, 1);
     apf::zeroField(strs);
@@ -68,84 +67,106 @@ namespace bio
     axl_yngs_mod =
         apf::createIPField(apf_mesh, "axial_youngs_modulus", apf::SCALAR, 1);
     apf::zeroField(axl_yngs_mod);
-    amsi::applyUniqueRegionTags(imdl, part, apf_mesh);
+    amsi::applyUniqueRegionTags(apf_mesh);
     // zero stiffness_variation and axial_youngs_modulus fields
-    std::vector<pANode> stf_vrtn_nds;
-    amsi::cutPaste<pANode>(
-        AttNode_childrenByType((pANode)pd, "stiffness gradient"),
-        std::back_inserter(stf_vrtn_nds));
-    for (auto nd = stf_vrtn_nds.begin(); nd != stf_vrtn_nds.end(); ++nd)
-      stf_vrtn_cnst.push_back(buildStiffnessVariation(pd, *nd, stf_vrtn));
-    for (auto cnst = stf_vrtn_cnst.begin(); cnst != stf_vrtn_cnst.end(); ++cnst)
-      (*cnst)->populate_stf_vrtn_fld();
-    //    Sim_logOff();
-    std::vector<pANode> vol_cnst_nds;
-    amsi::cutPaste<pANode>(AttNode_childrenByType((pANode)pd, "incompressible"),
-                           std::back_inserter(vol_cnst_nds));
-    for (auto vol_nd = vol_cnst_nds.begin(); vol_nd != vol_cnst_nds.end();
-         ++vol_nd)
-      vol_cnst.push_back(
-          buildVolumeConstraint(pd, *vol_nd, apf_primary_numbering));
-    pGEntity rgn = NULL;
-    GRIter ri = GM_regionIter(imdl);
-    while ((rgn = (pGEntity)GRIter_next(ri))) {
-      pAttribute mm = GEN_attrib(rgn, "material model");
-      assert(mm);
-      pAttribute cm = Attribute_childByType(mm, "continuum model");
-      assert(cm);
-      char* img_cls = Attribute_imageClass(cm);
-      int cnst_type = getConstitutiveTypeFromString(img_cls);
-      Sim_deleteString(img_cls);
-      if (cnst_type == isotropic_neohookean)  // push evaluation inside of
-                                              // elemental system constructor?
+    const auto * stiffness_gradient =
+        problem_definition.unassociated->FindCategoryNodeByType(
+            "stiffness gradient");
+    if (stiffness_gradient != nullptr)
+    {
+      for (const auto & grad_nd : stiffness_gradient->GetCategoryNodes())
+      {
+        stf_vrtn_cnst.push_back(
+            buildStiffnessVariation(*stiffness_gradient, stf_vrtn));
+      }
+    }
+    for (auto & cnst : stf_vrtn_cnst)
+    {
+      cnst->populate_stf_vrtn_fld();
+    }
+    const auto * incompressible =
+        problem_definition.unassociated->FindCategoryNodeByType(
+            "incompressible");
+    if (incompressible != nullptr)
+    {
+      for (const auto & incompressible_constraint :
+           incompressible->GetCategoryNodes())
+      {
+        vol_cnst.push_back(buildVolumeConstraint(incompressible_constraint,
+                                                 apf_primary_numbering));
+      }
+    }
+    static constexpr int dimension = 3;
+    auto * gmodel = mesh->getModel();
+    struct gmi_ent * gent;
+    auto * it = gmi_begin(gmodel, dimension);
+    while ((gent = gmi_next(gmodel, it)))
+    {
+      int tag = gmi_tag(gmodel, gent);
+      const auto * model_traits =
+          problem_definition.associated->Find({dimension, tag});
+      const auto * material_model =
+          mt::GetCategoryByType(model_traits, "material model");
+      const auto * continuum_model =
+          mt::GetPrimaryCategoryByType(model_traits, "continuum model");
+      const auto * youngs_modulus =
+          mt::GetCategoryModelTraitByType<mt::ScalarMT>(continuum_model,
+                                                        "youngs modulus");
+      const auto * poisson_ratio =
+          mt::GetCategoryModelTraitByType<mt::ScalarMT>(continuum_model,
+                                                        "poisson ratio");
+      if (youngs_modulus == nullptr || poisson_ratio == nullptr)
+      {
+        std::cerr << " \"poisson ratio\" and \"youngs modulus\" are required "
+                     "types for the continuum model.\n";
+        MPI_Abort(AMSI_COMM_WORLD, 1);
+      }
+      if (continuum_model->GetType() == "isotropic_neohookean")
       {
         // should check to make sure the continuum model is iso lin ela for init
         // solve?
-        pAttributeTensor0 yng =
-            (pAttributeTensor0)Attribute_childByType(cm, "youngs modulus");
-        pAttributeTensor0 psn =
-            (pAttributeTensor0)Attribute_childByType(cm, "poisson ratio");
-        double E = AttributeTensor0_value(yng);
-        double v = AttributeTensor0_value(psn);
-        constitutives[rgn] = new NeoHookeanIntegrator(this, apf_primary_field, dfm_grd, current_coords, E, v, 1);
+        constitutives[reinterpret_cast<apf::ModelEntity *>(gent)] =
+            std::make_unique<NeoHookeanIntegrator>(
+                this, apf_primary_field, dfm_grd, current_coords,
+                (*youngs_modulus)(), (*poisson_ratio)(), 1);
       }
-      else if (cnst_type == transverse_isotropic) {
-        pAttributeTensor0 yng =
-            (pAttributeTensor0)Attribute_childByType(cm, "youngs modulus");
-        pAttributeTensor0 psn =
-            (pAttributeTensor0)Attribute_childByType(cm, "poisson ratio");
-        pAttributeTensor1 axial_axs =
-            (pAttributeTensor1)Attribute_childByType(cm, "axis");
-        pAttributeTensor0 axial_shr =
-            (pAttributeTensor0)Attribute_childByType(cm, "axial shear modulus");
-        pAttributeTensor0 axial_ygn = (pAttributeTensor0)Attribute_childByType(
-            cm, "axial youngs modulus");
-        double E = AttributeTensor0_value(yng);
-        double v = AttributeTensor0_value(psn);
-        double axs[3] = {0.0, 0.0, 0.0};  // axis must be constant for now
-        AttributeTensor1_evalTensorDT(axial_axs, 0.0, &axs[0]);
-        double axialG = AttributeTensor0_value(axial_shr);
-        double axialE = AttributeTensor0_value(axial_ygn);
-        constitutives[rgn] = new TrnsIsoNeoHookeanIntegrator(
-	  this, apf_primary_field, stf_vrtn, dfm_grd, axl_yngs_mod, current_coords, E, v,
-            &axs[0], axialG, axialE, 1);
+      else if (continuum_model->GetType() == "transverse_isotropic")
+      {
+        const auto * axis = mt::GetCategoryModelTraitByType<mt::VectorMT>(
+            continuum_model, "axis");
+        const auto * axial_shear_modulus =
+            mt::GetCategoryModelTraitByType<mt::ScalarMT>(
+                continuum_model, "axial shear modulus");
+        const auto * axial_youngs_modulus =
+            mt::GetCategoryModelTraitByType<mt::ScalarMT>(
+                continuum_model, "axial youngs modulus");
+        std::array<double, 3> axs = {(*axis)(0), (*axis)(1), (*axis)(2)};
+        constitutives[reinterpret_cast<apf::ModelEntity *>(gent)] =
+            std::make_unique<TrnsIsoNeoHookeanIntegrator>(
+                this, apf_primary_field, stf_vrtn, dfm_grd, axl_yngs_mod,
+                current_coords, (*youngs_modulus)(), (*poisson_ratio)(),
+                &axs[0], (*axial_shear_modulus)(), (*axial_youngs_modulus)(),
+                1);
       }
     }
-    GRIter_delete(ri);
-    int dir_tps[] = {amsi::FieldUnit::displacement};
-    amsi::buildSimBCQueries(pd, amsi::BCType::dirichlet, &dir_tps[0],
-                            (&dir_tps[0]) + 1, std::back_inserter(dir_bcs));
-    int neu_tps[] = {amsi::NeuBCType::traction, amsi::NeuBCType::pressure};
-    amsi::buildSimBCQueries(pd, amsi::BCType::neumann, &neu_tps[0],
-                            (&neu_tps[0]) + 2, std::back_inserter(neu_bcs));
+    gmi_end(gmodel, it);
+    dirichlet_bcs.push_back(
+        amsi::DirichletBCEntry{.categories = {"displacement"}, .mt_name = "x"});
+    dirichlet_bcs.push_back(
+        amsi::DirichletBCEntry{.categories = {"displacement"}, .mt_name = "y"});
+    dirichlet_bcs.push_back(
+        amsi::DirichletBCEntry{.categories = {"displacement"}, .mt_name = "z"});
+    neumann_bcs.push_back(
+        amsi::NeumannBCEntry{.categories = {"pressure"},
+                             .mt_name = "magnitude",
+                             .mt_type = amsi::NeumannBCType::pressure});
+    neumann_bcs.push_back(
+        amsi::NeumannBCEntry{.categories = {"traction"},
+                             .mt_name = "direction",
+                             .mt_type = amsi::NeumannBCType::traction});
   }
   NonlinearTissue::~NonlinearTissue()
   {
-    // clean up constitutive properties
-    for(auto it=constitutives.begin(); it!=constitutives.end(); ++it)
-    {
-      delete it->second;
-    }
     delete xpyfnc;
     apf::destroyField(current_coords);
     delete prv_crd_fnc;
@@ -153,15 +174,17 @@ namespace bio
     // destroying this field currently throws an error
     // most likely because it is zeroed and core does not
     // properly deal with this...
-    //apf::destroyField(delta_u);
+    // apf::destroyField(delta_u);
     apf::destroyField(strs);
     apf::destroyField(strn);
     apf::destroyField(stf_vrtn);
     apf::destroyField(axl_yngs_mod);
   }
-  void NonlinearTissue::computeInitGuess(amsi::LAS* las)
+  void NonlinearTissue::computeInitGuess(amsi::LAS * las)
   {
-    LinearTissue lt(model, mesh, prob_def, solution_strategy, analysis_comm);
+    // LinearTissue lt(model, mesh, prob_def, solution_strategy, analysis_comm);
+    LinearTissue lt(apf_mesh, problem_definition, solution_strategy, output,
+                    analysis_comm);
     lt.setSimulationTime(T);
     LinearSolver(&lt, las);
     las->iter();
@@ -181,18 +204,19 @@ namespace bio
       (*cnst)->iter();
     iteration++;
   }
-  void NonlinearTissue::Assemble(amsi::LAS* las)
+  void NonlinearTissue::Assemble(amsi::LAS * las)
   {
     ApplyBC_Neumann(las);
-    apf::MeshEntity* me = NULL;
+    apf::MeshEntity * me = NULL;
     // custom iterator would be perfect for switching for multiscale version
     auto it = apf_mesh->begin(analysis_dim);
     // FIXME shouldn't we skip non-owned elements similar to
     // apf::integrator::process(apf::Mesh*)?
-    while ((me = apf_mesh->iterate(it))) {
-      amsi::ElementalSystem* constitutive =
-          constitutives[R_whatIn((pRegion)me)];
-      apf::MeshElement* mlmt = apf::createMeshElement(apf_mesh, me);
+    while ((me = apf_mesh->iterate(it)))
+    {
+      amsi::ElementalSystem * constitutive =
+          constitutives[apf_mesh->toModel(me)].get();
+      apf::MeshElement * mlmt = apf::createMeshElement(apf_mesh, me);
       apf::Element * elm = apf::createElement(constitutive->getField(), mlmt);
       // amsi::ElementalSystem * sys = getElementalSystem(me,0); // assumes 1
       // type of system per element
@@ -213,9 +237,9 @@ namespace bio
     // process constraints
     for (auto cnst = vol_cnst.begin(); cnst != vol_cnst.end(); cnst++)
       (*cnst)->apply(las);
-    //las->GetVectorNorm(nrm);
+    // las->GetVectorNorm(nrm);
   }
-  void NonlinearTissue::UpdateDOFs(const double* sol)
+  void NonlinearTissue::UpdateDOFs(const double * sol)
   {
     // accumulate displacement deltas into primary field
     // apfSimFEA::UpdateDOFs(sol);
@@ -234,13 +258,65 @@ namespace bio
     apf::synchronize(apf_primary_field);
     apf::synchronize(delta_u);
   }
-  void NonlinearTissue::getLoadOn(pGEntity ent, double* frc)
+  /**
+   * get the Neumann boundary condition value on the specified entity
+   * @param ent model entity to query
+   * @param frc will be filled with the value of the NeumannBCs
+   */
+  void NonlinearTissue::getLoadOn(apf::ModelEntity * ent, double * frc)
   {
-    amsi::SimBCQuery* snbcq =
-        amsi::findSimBCQueryOn(neu_bcs.begin(), neu_bcs.end(), ent);
-    assert(snbcq);
-    assert(!snbcq->isSpaceExpr());
-    for (int ii = 0; ii < 3; ii++) frc[ii] = snbcq->getValue(ii, T);
+    auto model_dimension = apf_mesh->getModelType(ent);
+    auto model_tag = apf_mesh->getModelTag(ent);
+    auto * associated_traits = problem_definition.associated->Find({model_dimension, model_tag});
+    if (associated_traits == nullptr)
+    {
+      std::cerr << "Attempting to get Neumann BC on [" << model_dimension << ","
+                << model_tag << "].";
+      std::cerr << " No boundary conditions associated with that geometry.\n";
+      exit(1);
+    }
+    for (const auto & neumann_bc : neumann_bcs)
+    {
+      const mt::AssociatedCategoryNode * category_node = associated_traits;
+      for (const auto & category : neumann_bc.categories)
+      {
+        category_node = category_node->FindCategoryByType(category);
+        if (category_node == nullptr)
+        {
+          break;
+        }
+      }
+      if (category_node == nullptr)
+      {
+        std::cerr << "Invalid Neumann BC provided\n";
+        exit(1);
+      }
+      const auto * load_trait =
+          category_node->FindModelTrait(neumann_bc.mt_name);
+      const auto * const_load_trait = mt::MTCast<mt::VectorMT>(load_trait);
+      const auto * time_function_load =
+          mt::MTCast<mt::VectorFunctionMT<1>>(load_trait);
+      if (const_load_trait != nullptr)
+      {
+        for (int i = 0; i < 3; ++i)
+        {
+          frc[i] = (*const_load_trait)(i);
+        }
+      }
+      else if (time_function_load != nullptr)
+      {
+        for (int i = 0; i < 3; ++i)
+        {
+          frc[i] = (*time_function_load)(i, T);
+        }
+      }
+      else
+      {
+        std::cerr << "Load must be either a scalar value, or function of time "
+                     "only.\n";
+        exit(1);
+      }
+    }
   }
   void NonlinearTissue::recoverSecondaryVariables(int load_step)
   {
@@ -259,28 +335,28 @@ namespace bio
     */
     //#endif
   }
-  void NonlinearTissue::storeStrain(apf::MeshElement* me, double* strain)
+  void NonlinearTissue::storeStrain(apf::MeshElement * me, double * strain)
   {
-    apf::MeshEntity* m_ent = apf::getMeshEntity(me);
+    apf::MeshEntity * m_ent = apf::getMeshEntity(me);
     apf::Matrix3x3 eps(strain[0], strain[3], strain[5], strain[3], strain[1],
                        strain[4], strain[5], strain[4], strain[2]);
     apf::setMatrix(strn, m_ent, 0, eps);
   }
-  void NonlinearTissue::storeStrain(apf::MeshElement* me, apf::Matrix3x3 eps)
+  void NonlinearTissue::storeStrain(apf::MeshElement * me, apf::Matrix3x3 eps)
   {
-    apf::MeshEntity* m_ent = apf::getMeshEntity(me);
+    apf::MeshEntity * m_ent = apf::getMeshEntity(me);
     apf::setMatrix(strn, m_ent, 0, eps);
   }
-  void NonlinearTissue::storeStress(apf::MeshElement* me, double* stress)
+  void NonlinearTissue::storeStress(apf::MeshElement * me, double * stress)
   {
-    apf::MeshEntity* m_ent = apf::getMeshEntity(me);
+    apf::MeshEntity * m_ent = apf::getMeshEntity(me);
     apf::Matrix3x3 sigma(stress[0], stress[3], stress[5], stress[3], stress[1],
                          stress[4], stress[5], stress[4], stress[2]);
     apf::setMatrix(strs, m_ent, 0, sigma);
   }
-  void NonlinearTissue::storeStress(apf::MeshElement* me, apf::Matrix3x3 eps)
+  void NonlinearTissue::storeStress(apf::MeshElement * me, apf::Matrix3x3 eps)
   {
-    apf::MeshEntity* m_ent = apf::getMeshEntity(me);
+    apf::MeshEntity * m_ent = apf::getMeshEntity(me);
     apf::setMatrix(strs, m_ent, 0, eps);
   }
-}
+}  // namespace bio
