@@ -1,23 +1,30 @@
+#include <amsiNonlinearConvergenceOperator.h>
+#include <model_traits/AssociatedModelTraits.h>
+#include <model_traits/ModelTraits.h>
 #include "bioVolumeConvergence.h"
-#include <simNonlinearAnalysis.h>
 namespace bio
 {
-  /**
-   * Extracts the volume convergence regions from a simmetrix case
-   * @param ss simmetrix case
-   * @param las the amsi las system used for convergence
-   * @param out typically std::back_inserter to some sort of list type.
-   */
   template <typename O>
-    void buildLASConvergenceOperators(pACase ss, amsi::MultiIteration * it, amsi::LAS * las, O out)
+  void buildLASConvergenceOperators(const mt::CategoryNode * solution_strategy,
+                                    amsi::MultiIteration * it,
+                                    amsi::LAS * las,
+                                    O out)
   {
-    std::vector<pAttribute> cnvrg_atts;
-    amsi::cutPaste<pAttribute>(AttCase_attributes(ss,"convergence operator"),std::back_inserter(cnvrg_atts));
-    for(auto cnvrg_att = cnvrg_atts.begin(); cnvrg_att != cnvrg_atts.end(); ++cnvrg_att)
+    auto convergence_operators_cat =
+        mt::GetCategoriesByType(solution_strategy, "convergence operator");
+    for (const auto * convergence_operator_cat : convergence_operators_cat)
     {
-      auto cvg = amsi::buildSimConvergenceOperator(ss,*cnvrg_att,it,las);
-      if(cvg != NULL)
-        *out++ = cvg;
+      for (const auto & convergence_operator :
+           convergence_operator_cat->GetCategoryNodes())
+      {
+        auto * cvg =
+            amsi::createConvergenceOperator(&convergence_operator, it, las)
+                .release();
+        if (cvg != nullptr)
+        {
+          *out++ = cvg;
+        }
+      }
     }
   }
   /**
@@ -29,20 +36,37 @@ namespace bio
    * @param out typically std::back_inserter to some sort of list type.
    */
   template <typename I, typename O>
-    void buildVolConvergenceOperators(pACase ss, amsi::MultiIteration * it, apf::Field * u, I vl_tks, O out)
+  void buildVolConvergenceOperators(const mt::CategoryNode * solution_strategy,
+                                    amsi::MultiIteration * it,
+                                    apf::Field * u,
+                                    I vl_tks,
+                                    O out)
   {
-    std::vector<pAttribute> cnvrg_atts;
-    amsi::cutPaste<pAttribute>(AttCase_attributes(ss,"convergence operator"),std::back_inserter(cnvrg_atts));
-    for(auto cnvrg_att = cnvrg_atts.begin(); cnvrg_att != cnvrg_atts.end(); ++cnvrg_att)
+    const auto * convergence_operator =
+        mt::GetCategoryByType(solution_strategy, "convergence operator");
+    auto volume_convergence_operators =
+        mt::GetCategoriesByType(convergence_operator, "volume convergence");
+    for (const auto * volume_convergence : volume_convergence_operators)
     {
-      char * tp = Attribute_imageClass(*cnvrg_att);
-      if(std::string("volume convergence").compare(tp) == 0)
+      const auto * regions =
+          mt::GetCategoryByType(volume_convergence, "regions");
+      const auto * track_volume =
+          mt::GetCategoryModelTraitNodeByType(regions, "track volume");
+      if (track_volume == nullptr)
       {
-        pAttribute rgn_nd_att = Attribute_childByType(*cnvrg_att,"regions");
-        pANode rgn_nd = AttributeRefNode_value((pAttributeRefNode)rgn_nd_att);
-        *out++ = buildVolConvergenceOperator(ss,*cnvrg_att,it,vl_tks[rgn_nd],u);
+        std::cerr << "track volume required on volume convergence.\n";
+        MPI_Abort(AMSI_COMM_WORLD, 1);
       }
-      Sim_deleteString(tp);
+      const auto & model_traits = track_volume->GetModelTraits();
+      if (model_traits.size() != 1 || track_volume->GetName().empty())
+      {
+        std::cerr << "track volume region should only have one model trait and "
+                     "must have a name!\n";
+        MPI_Abort(AMSI_COMM_WORLD, 1);
+      }
+      *out++ = buildVolConvergenceOperator(*volume_convergence, it,
+                                           vl_tks[track_volume->GetName()], u)
+                   .release();
     }
   }
-}
+}  // namespace bio
