@@ -1,33 +1,42 @@
 #ifndef BIO_STIFFNESS_VARIATION_H_
 #define BIO_STIFFNESS_VARIATION_H_
-#include <simAnalysis.h>
-#include <apfSIM.h>
-#include <cstring>
 #include <apf.h>
+#include <model_traits/CategoryNode.h>
+#include <model_traits/ModelTrait.h>
+#include <apfMesh.h>
+#include <cstring>
+#include <memory>
 namespace bio
 {
   class StiffnessVariation;
-  StiffnessVariation * buildStiffnessVariation(pACase cs, pANode nd, apf::Field * stf_vrtn);
+  std::unique_ptr<StiffnessVariation> buildStiffnessVariation(
+      const mt::CategoryNode & nd,
+      apf::Field * stf_vrtn_fld);
   class StiffnessVariation : public apf::Integrator
   {
-  protected:
+    protected:
     apf::Field * stf_vrtn_fld;
-    std::vector<apf::ModelEntity*> mdl_src_ents;
-    std::vector<apf::ModelEntity*> mdl_snk_ents;
-    double * target_xyz;
+    std::vector<apf::ModelEntity *> mdl_src_ents;
+    std::vector<apf::ModelEntity *> mdl_snk_ents;
+    apf::Vector3 target_xyz;
     apf::MeshEntity * msh_ent;
     int ip_integration_pt;
-    pAttributeDouble fn;
-  public:
-    template <typename I>
-      StiffnessVariation(I mdl_src_ent_bgn, I mdl_src_ent_end, I mdl_snk_ent_bgn, I mdl_snk_end_end, apf::Field * stf_vrtn_fld, pAttributeDouble fn);
-    void inElement(apf::MeshElement * msh_lmnt){msh_ent = apf::getMeshEntity(msh_lmnt); ip_integration_pt=0;}
-    void outElement(){};
-    void atPoint(apf::Vector3 const &, double, double);
-    void set_target_xyz(double * xyz){target_xyz = xyz;}
+    std::shared_ptr<const mt::ScalarFunctionMT<1>> fn;
+    public:
+    StiffnessVariation(std::vector<apf::ModelEntity*> src_ents,
+                       std::vector<apf::ModelEntity*> sink_ents,
+                       apf::Field * stf_vrtn_fld,
+                       std::shared_ptr<const mt::ScalarFunctionMT<1>> fn);
+    void inElement(apf::MeshElement * msh_lmnt) final
+    {
+      msh_ent = apf::getMeshEntity(msh_lmnt);
+      ip_integration_pt = 0;
+    }
+    void outElement() final {};
+    void atPoint(apf::Vector3 const &, double, double) final;
+    void set_target_xyz(const apf::Vector3 & xyz) { target_xyz = xyz; }
     void populate_stf_vrtn_fld();
-    void calculate_mdl_fc_coord(apf::ModelEntity * mdl_fc, double * mdl_fc_xyz);
+    void calculate_mdl_fc_coord(apf::ModelEntity * mdl_fc, apf::Vector3& mdl_fc_xyz);
   };
-}
-#include "bioStiffnessVariation_impl.h"
+}  // namespace bio
 #endif
